refactor(Authentication): Refactored Authentication to external modules, work on typedoc generation
