+++ package.json
@@ -14,7 +14,9 @@
     "semantic-release": "semantic-release pre && semantic-release post",
     "lint": "tslint src/**/*.ts --project tsconfig.json",
     "prebuild": "gulp build:clean",
-    "build": "tsc"
+    "build": "tsc",
+    "typedoc:md": "gulp typedoc:md",
+    "typedoc:html": "gulp typedoc"
   },
   "repository": {
     "type": "git",
@@ -42,7 +44,8 @@
     "app-root-path": "2.0.1",
     "gulp-run": "^1.7.1",
     "nyc": "^10.2.0",
-    "ts-json-properties": "1.1.4"
+    "ts-json-properties": "1.1.4",
+    "typedoc-plugin-external-module-name": "^1.0.9"
   },
   "devDependencies": {
     "@types/app-root-path": "1.2.4",
@@ -57,6 +60,7 @@
     "fs-then-native": "2.0.0",
     "gulp": "3.9.1",
     "gulp-header": "1.8.8",
+    "gulp-rename": "^1.2.2",
     "gulp-standard": "9.0.0",
     "gulp-tslint": "7.1.0",
     "gulp-typedoc": "2.0.2",
@@ -68,6 +72,7 @@
     "ts-json-properties": "1.1.4",
     "tslint": "4.5.1",
     "typedoc": "0.5.10",
+    "typedoc-md-theme": "^1.0.1",
     "typescript": "2.3.0"
   },
   "czConfig": {
