+++ src/Authentication/Authentication.ts
@@ -1,148 +1,12 @@
-import { Repository } from './SN';
-import { Subject, BehaviorSubject } from '@reactivex/rxjs';
-import { Observable } from '@reactivex/rxjs';
+/**
+ * @module "Authentication"
+ * 
+ * This module is a library module that contains authentication-related classes and references
+ */ /** */
 
-export enum LoginState {
-    Pending,
-    Unauthenticated,
-    Authenticated
-}
-
-type TokenType = 'access' | 'refresh';
-
-class LoginResponse {
-    access: string;
-    refresh: string;
-}
-
-class RefreshResponse {
-    access: string;
-}
-
-interface ITokenPayload {
-    iss: string;
-    sub: string;
-    aud: string;
-    exp: number;
-    iat: number;
-    nbf: number;
-    name: string;
-}
-
-class Token {
-
-    private get tokenPayload(): ITokenPayload {
-        try {
-            return JSON.parse(Buffer.from(this.payloadEncoded, 'base64').toString()) as ITokenPayload;
-        } catch (err) {
-            return {
-                aud: '',
-                exp: 0,
-                iat: 0,
-                iss: '',
-                name: '',
-                nbf: 0,
-                sub: ''
-            }
-        }
-    };
-
-    private fromEpoch(epoch: number): Date {
-        let d = new Date(0);
-        d.setUTCSeconds(epoch);
-        return d;
-    }
-
-    public get Username(): string {
-        return this.tokenPayload.name;
-    };
-
-    public GetPayload() {
-        return this.tokenPayload;
-    }
-
-    public get ExpirationTime() {
-        return this.fromEpoch(this.tokenPayload.exp);
-    }
-
-    public get NotBefore() {
-        return this.fromEpoch(this.tokenPayload.nbf);
-    }
-
-    public IsValid() {
-        let now = new Date();
-        return this.tokenPayload && this.ExpirationTime > now && this.NotBefore < now;
-    }
-
-    public get IssuedDate() {
-        return this.fromEpoch(this.tokenPayload.iat);
-    }
-
-    public toString() {
-        return `${this.headerEncoded}.${this.payloadEncoded}`;
-    }
-
-
-    public static FromHeadAndPayload(headAndPayload: string): Token {
-        let [head, payload] = headAndPayload.split('.');
-        return new Token(head, payload)
-    }
-
-    public static get Empty(): Token {
-        return new Token('', '');
-    }
-
-    private constructor(private readonly headerEncoded: string, private readonly payloadEncoded: string) {
-    }
-}
-
-class TokenStore {
-
-    constructor(private readonly baseUrl: string) {
-    }
-
-    private get tokenKeyPrefix() {
-        return `sn-${this.baseUrl}-`;
-    }
-
-    private innerStore: string[] = [];
-    public GetToken(key: TokenType): Token {
-        const storeKey = `${this.tokenKeyPrefix}${key}`;
-        try {
-            if (typeof localStorage === 'undefined') {
-                return Token.FromHeadAndPayload(this.innerStore[storeKey]);
-            }
-            return Token.FromHeadAndPayload(localStorage.getItem(storeKey));
-        } catch (err) {
-            return Token.Empty;
-        }
-    }
-
-    public SetToken(key: TokenType, token: Token) {
-        const storeKey = `${this.tokenKeyPrefix}${key}`;
-        let dtaString = token.toString();
-        if (typeof localStorage === 'undefined') {
-            this.innerStore[storeKey] = dtaString;
-        } else {
-            localStorage.setItem(storeKey, dtaString);
-        }
-    }
-
-    public get AccessToken() {
-        return this.GetToken('access');
-    }
-    public set AccessToken(value: Token) {
-        this.SetToken('access', value);
-    }
-
-    public get RefreshToken() {
-        return this.GetToken('refresh');
-    }
-    public set RefreshToken(value: Token) {
-        this.SetToken('refresh', value);
-    }
-
-}
+import { LoginState, LoginResponse, RefreshResponse, ITokenPayload, Token, TokenStore } from './';
+import { Subject, BehaviorSubject, Observable } from '@reactivex/rxjs';
+import { Repository } from '../SN';
 
 export class Authentication {
     public readonly State: BehaviorSubject<LoginState> = new BehaviorSubject<LoginState>(LoginState.Pending);
