+++ src/Authentication/ITokenPayload.ts
@@ -1,0 +1,12 @@
+/**
+ * @module "Authentication"
+ */ /** */
+export interface ITokenPayload {
+    iss: string;
+    sub: string;
+    aud: string;
+    exp: number;
+    iat: number;
+    nbf: number;
+    name: string;
+}
