+++ src/Authentication/LoginResponse.ts
@@ -1,0 +1,7 @@
+/**
+ * @module "Authentication"
+ */ /** */
+export class LoginResponse {
+    access: string;
+    refresh: string;
+}
