+++ src/Authentication/LoginState.ts
@@ -1,0 +1,8 @@
+/**
+ * @module "Authentication"
+ */ /** */
+export enum LoginState {
+    Pending,
+    Unauthenticated,
+    Authenticated
+}
