+++ src/Authentication/RefreshResponse.ts
@@ -1,0 +1,6 @@
+/**
+ * @module "Authentication"
+ */ /** */
+export class RefreshResponse {
+    access: string;
+}
